IT WORKS (kinda)

+++ crates/text/src/selection.rs
@@ -6,7 +6,7 @@ use std::ops::Range;
 pub enum SelectionGoal {
     None,
     HorizontalPosition(f64),
-    HorizontalRange { start: f32, end: f32 },
+    HorizontalRange { start: f64, end: f64 },
     WrappedHorizontalPosition((u32, f32)),
 }
