+++ crates/vim/src/motion.rs
@@ -1573,8 +1573,8 @@ fn up_down_buffer_rows(
 
     let (goal_wrap, goal_x) = match goal {
         SelectionGoal::WrappedHorizontalPosition((row, x)) => (row, x),
-        SelectionGoal::HorizontalRange { end, .. } => (select_nth_wrapped_row, end),
-        SelectionGoal::HorizontalPosition(x) => (select_nth_wrapped_row, x),
+        SelectionGoal::HorizontalRange { end, .. } => (select_nth_wrapped_row, end as f32),
+        SelectionGoal::HorizontalPosition(x) => (select_nth_wrapped_row, x as f32),
         _ => {
             let x = map.x_for_display_point(point, text_layout_details);
             goal = SelectionGoal::WrappedHorizontalPosition((select_nth_wrapped_row, x.into()));
